% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_utils.R
\name{construct_payload}
\alias{construct_payload}
\title{Construct Payload for Loki}
\usage{
construct_payload(log_message, log_labels, trace)
}
\arguments{
\item{log_message}{The log message as a character string.}

\item{log_labels}{The labels for the log as a list.}

\item{trace}{Traceback information as a character vector or NULL.}
}
\value{
A list representing the payload.
}
\description{
Constructs the payload to send to Loki.
}
\keyword{internal}
