% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_utils.R
\name{get_loki_endpoint}
\alias{get_loki_endpoint}
\title{Get Loki Endpoint}
\usage{
get_loki_endpoint(loki_endpoint, pkg_env)
}
\arguments{
\item{loki_endpoint}{The endpoint URL as a string or NULL.}

\item{pkg_env}{The package environment containing global settings.}
}
\value{
The Loki endpoint as a string.
}
\description{
Gets the Loki endpoint either from the function argument or the package environment.
}
\keyword{internal}
