% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_to_loki}
\alias{log_to_loki}
\title{Log Messages to Loki with Custom Labels}
\usage{
log_to_loki(
  log_message,
  log_labels,
  loki_endpoint = NULL,
  trace = NULL,
  async = TRUE
)
}
\arguments{
\item{log_message}{Character string specifying the log message.}

\item{log_labels}{Named list specifying the labels you want to add to the log.}

\item{loki_endpoint}{Character string specifying the URL of the Loki instance to send logs to.}

\item{trace}{Optional, character vector containing the traceback information. Default is NULL.}

\item{async}{Logical indicating whether to send the log asynchronously. Default is TRUE.}
}
\value{
A list containing the HTTP response information from Loki, or a future promise if async is TRUE.
}
\description{
This function allows you to send log messages from R to a Loki instance with custom labels.
}
\examples{
\dontrun{
log_to_loki("This is a test log", list(app = "MyApp", level = "INFO"), "http://your-loki-instance:3100/loki/api/v1/push")
log_to_loki("This is an error", list(app = "MyApp", level = "ERROR"), "http://your-loki-instance:3100/loki/api/v1/push", trace = traceback(2, as.character = TRUE))
}
}
