% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_utils.R
\name{send_log_to_loki}
\alias{send_log_to_loki}
\title{Send Log to Loki}
\usage{
send_log_to_loki(json_payload, loki_endpoint, async = FALSE)
}
\arguments{
\item{json_payload}{The JSON payload as a string.}

\item{loki_endpoint}{The Loki endpoint as a string.}

\item{async}{Logical indicating whether to send the log asynchronously.}
}
\value{
Logical indicating whether the log was successfully sent, or a future promise.
}
\description{
Sends the JSON payload to Loki and checks the response.
}
\keyword{internal}
