% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_loki_endpoint.R
\name{set_loki_endpoint}
\alias{set_loki_endpoint}
\title{Set Loki Endpoint for rLoggerLoki}
\usage{
set_loki_endpoint(endpoint)
}
\arguments{
\item{endpoint}{A character string specifying the URL of the Loki endpoint.
This should include the protocol (usually 'http') and the domain name or IP address.
Optionally, the port number can also be included.}
}
\description{
This function sets the Loki endpoint that will be used for logging by the
\code{\link{log_to_loki}} function when no endpoint is explicitly provided.
}
\examples{
\dontrun{
set_loki_endpoint("http://your-loki-instance:3100")
}
}
