% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_utils.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate Log Arguments}
\usage{
validate_arguments(log_message, log_labels)
}
\arguments{
\item{log_message}{The log message as a character string.}

\item{log_labels}{The labels for the log as a list.}
}
\description{
Validates the arguments for logging.
}
\keyword{internal}
